Move adapter tests
