+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/config/report/BenchmarkReportConfig.java
@@ -33,7 +33,7 @@ import org.optaplanner.benchmark.impl.result.PlannerBenchmarkResult;
 import org.optaplanner.benchmark.impl.result.SolverBenchmarkResult;
 import org.optaplanner.core.config.AbstractConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
-import org.optaplanner.core.impl.io.jaxb.adapters.JaxbLocaleAdapter;
+import org.optaplanner.core.impl.io.jaxb.adapter.JaxbLocaleAdapter;
 
 @XmlType(propOrder = {
         "locale",
