+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/io/jaxb/adapter/JaxbDurationAdapterTest.java
@@ -1,12 +1,12 @@
 /*
  * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
+ *  
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ *  
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.io.jaxb;
+package org.optaplanner.core.impl.io.jaxb.adapter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.Duration;
 
 import org.junit.jupiter.api.Test;
-import org.optaplanner.core.impl.io.jaxb.adapters.JaxbDurationAdapter;
 
 public class JaxbDurationAdapterTest {
